EVERYTHING IS TIED TOGETHER - Map is now connected to shop and combat separately - New game and load game now lead to the game map - All required files are supposedly in Game folder

+++ Game/login.py
@@ -1,6 +1,6 @@
 import sqlite3
 from sqlite3 import Error
-from flask import Flask, render_template, request, flash, redirect, url_for, session, json
+from flask import Flask, render_template, request, flash, redirect, url_for, session, json, jsonify
 from flask_session import Session
 from flask_bcrypt import Bcrypt, check_password_hash, generate_password_hash
 import os
@@ -115,45 +115,49 @@ def login():
 def main_menu():
     return render_template("mainMenu.html")
 
-@app.route('/load-game')
+@app.route('/load-game', methods = ['POST'])
 def load_user():
-    print(current_user)
-    loadPlayer(current_user)
-    print(current_player_data['moves'])
-    if (len(current_player_data['moves']) == 0):
-        flash('You haven\'t started a game yet!')
-        return redirect(url_for('main_menu'))
-    return redirect(url_for('combat'))
+    if request.method == 'POST':
+        print(current_user)
+        loadPlayer(current_user)
+        print(current_player_data['moves'])
+        if (len(current_player_data['moves']) == 0):
+            flash('You haven\'t started a game yet!')
+            return redirect(url_for('main_menu'))
+        return redirect('/start-map')
+    return redirect('/')
 
-@app.route('/new-game')
+@app.route('/new-game', methods = ['POST'])
 def new_game():
-    global starting_stats
-    global starting_moves
-    conn_users = sqlite3.connect('user_info.db')
-    #Deleting any data that's there
-    #try:
-    conn_users.execute("DELETE FROM user_moves WHERE user = ?", (current_user,))
-    conn_users.commit()
-    #except Error as e:
-    #    print(f"No data from user_moves to delete")
-    #    conn_users.rollback()
-    try:
-        conn_users.execute("DELETE FROM user_inventory WHERE user = ?", (current_user,))
+    if request.method == 'POST':
+        global starting_stats
+        global starting_moves
+        conn_users = sqlite3.connect('user_info.db')
+        #Deleting any data that's there
+        #try:
+        conn_users.execute("DELETE FROM user_moves WHERE user = ?", (current_user,))
         conn_users.commit()
-    except Error as e:
-        print(f"No data from user_inventory to delete")
-        conn_users.rollback()
-        
-    #Adding starting moves and setting player data to default, because they shouldn't have anything
-    conn_users.execute('UPDATE logins SET player_health = ?, player_atk = ?, player_def = ?, player_currency = ?, player_kills = ?, player_deaths = ?, player_waveflag = ?'
-                       + 'WHERE username = ?', (starting_stats[0], starting_stats[1], starting_stats[2], starting_stats[3], starting_stats[4], starting_stats[5], starting_stats[6], current_user))
-    for i in range(len(starting_moves)):
-        conn_users.execute("INSERT INTO user_moves (user, move_name) VALUES (?, ?)",(current_user, starting_moves[i]))
-    conn_users.commit()
-    conn_users.close()
-    
-    loadPlayer(current_user)
-    return redirect(url_for('combat'))
+        #except Error as e:
+        #    print(f"No data from user_moves to delete")
+        #    conn_users.rollback()
+        try:
+            conn_users.execute("DELETE FROM user_inventory WHERE user = ?", (current_user,))
+            conn_users.commit()
+        except Error as e:
+            print(f"No data from user_inventory to delete")
+            conn_users.rollback()
+            
+        #Adding starting moves and setting player data to default, because they shouldn't have anything
+        conn_users.execute('UPDATE logins SET player_health = ?, player_atk = ?, player_def = ?, player_currency = ?, player_kills = ?, player_deaths = ?, player_waveflag = ?'
+                            + 'WHERE username = ?', (starting_stats[0], starting_stats[1], starting_stats[2], starting_stats[3], starting_stats[4], starting_stats[5], starting_stats[6], current_user))
+        for i in range(len(starting_moves)):
+            conn_users.execute("INSERT INTO user_moves (user, move_name) VALUES (?, ?)",(current_user, starting_moves[i]))
+        conn_users.commit()
+        conn_users.close()
+
+        loadPlayer(current_user)
+        return redirect('/start-map')
+    return redirect('/')
 
 
 #----------------------------------------------------------------------------------
@@ -191,6 +195,140 @@ def clear_session():
     session.pop('_flashes',None)
     return redirect(url_for('login'))
 current_player_data = {}
+
+
+#----------------------------------------------------------------------------
+# MAP
+#----------------------------------------------------------------------------
+@app.route('/start-map')
+def startMap():
+    return redirect('/gameMap')
+
+@app.route('/gameMap')
+def mapRender():
+    loadPlayer(current_user)
+    itemNames = []
+    itemQuantities = []
+    for itemName, itemQuantity in current_player_data['inventory'].items():
+        itemNames.append(itemName)
+        itemQuantities.append(itemQuantity)
+
+    return render_template('map.html', hp = current_player_data['hp'], currency = current_player_data['currency'], 
+                           moves = current_player_data['moves'], itemNames = itemNames, itemQuantities = itemQuantities, len=len(itemNames))
+
+@app.route('/start-battle', methods=['POST'])
+def battle():
+    return redirect('/combat')
+
+@app.route('/start-shop', methods=['POST'])
+def start_shop():
+    return redirect('/shop')
+
+#----------------------------------------------------------------------------
+# SHOP
+#----------------------------------------------------------------------------
+@app.route("/shop")
+def display():
+    conn = sqlite3.connect('user_info.db')
+    cursor = conn.cursor()
+    username = current_user
+    # Assuming 'image_id' is the name of the item
+    cursor.execute(
+        'SELECT player_currency FROM logins WHERE username = ?',
+        (username,))
+    result = cursor.fetchone()
+    if result:
+        player_currency = result[0]
+    else:
+        player_currency = 0
+    conn.close()
+    return render_template('shop.html',currency=player_currency)
+
+def get_item_info_from_db(image_id):
+    conn = sqlite3.connect('items.db')
+    cursor = conn.cursor()
+    # Assuming 'image_id' is the name of the item
+    cursor.execute('SELECT description, hp_buff, atk_buff, def_buff, duration, cooldown, price FROM shop_items WHERE name = ?', (image_id,))
+
+    item = cursor.fetchone()
+    conn.close()
+    return item
+@app.route('/get_item_info', methods=['GET'])
+def get_item_info():
+    print("hola")
+    image_id = request.args.get('image_id')  # Assuming image ID is passed as a query parameter
+    if image_id is None:
+        return jsonify({'error': 'Image ID not provided'}), 400
+
+    print(image_id)
+    item = get_item_info_from_db(image_id)
+    if item is None:
+        return jsonify({'error': 'Item not found'}), 404
+
+    print("name")
+    # Assuming the database schema matches the returned columns
+    description, hp_buff, atk_buff, def_buff, duration, cooldown, price = item
+    print(description)
+
+    # Return item information as JSON response
+    return jsonify({
+        # 'name': name,
+        'description': description,
+        'hp_buff': hp_buff,
+        'atk_buff': atk_buff,
+        'def_buff': def_buff,
+        'duration': duration,
+        'cooldown': cooldown,
+        'price': price
+    })
+
+@app.route('/update_currency', methods=['POST'])
+def updateData():
+    try:
+        # Extract JSON data from request
+        req_data = request.get_json()
+        if req_data and 'currency' in req_data:
+            currency_value = req_data['currency']
+            conn = sqlite3.connect('user_info.db')
+            cursor = conn.cursor()
+            username = current_user
+            # Assuming 'image_id' is the name of the item
+            cursor.execute(
+                'UPDATE logins SET player_currency = ? WHERE username = ?',
+                (currency_value,username,))
+            conn.close()
+
+            image_id = req_data['item']
+            print("image_id")
+            conn = sqlite3.connect('user_info.db')
+            cursor = conn.cursor()
+            query = 'SELECT * FROM user_inventory WHERE user = ? AND item_name = ?'
+            cursor.execute(query, (username,image_id))
+            result = cursor.fetchone()  # Fetch one row (if exists)
+
+            if result is None:
+                cursor.execute('INSERT INTO user_inventory VALUES (?, ?, ?)',
+                               (username, image_id, 1))
+                conn.commit()
+                print("not in db")
+            else:
+                item = conn.execute('SELECT * FROM user_inventory WHERE user = ? AND item_name = ?',
+                                    (username, image_id)).fetchall()
+                print("newAmount")
+                newAmount = item[0][2] + 1
+                cursor.execute('UPDATE user_inventory SET quantity = ? WHERE user = ? AND item_name = ?',
+                             (newAmount, username, image_id))
+                conn.commit()
+                print("in db")
+            conn.close()
+            return jsonify({'success': True, 'currency': currency_value}), 200
+        else:
+            raise ValueError('Invalid JSON data')
+    except Exception as e:
+        error_msg = str(e)
+        print('Error:', error_msg)
+        return jsonify({'error': error_msg}), 400
+
 
 #----------------------------------------------------------------------------
 # COMBAT
@@ -300,7 +438,7 @@ def returnToMap():
         conn2.close()
         conn3.close()
         
-    return redirect(url_for('main_menu'))
+    return redirect("/start-map")
 
 def loadPlayer(user):
     conn = sqlite3.connect('user_info.db')
